Refactor to Swift 6
